Refactored code to validate the request body against the schema.
